requirements.txt
torch
torchvision
flask
pillow
opencv-python
numpy
transformers
